requirements.txt
Flask==3.1.3
flask_cors==6.0.2
mysql_connector_repackaged==0.3.1
python-dotenv==1.2.2
